--- .storybook/parts/wp-styles.js
@@ -1,8 +1,0 @@
-export const storybookWpStyles = () => {
-  require('EightshiftBlocksStorybookWp/editor/build-style/style-rtl.css');
-  require('EightshiftBlocksStorybookWp/editor/build-style/style.css');
-  require('EightshiftBlocksStorybookWp/components/build-style/style.css');
-  require('EightshiftBlocksStorybookWp/block-editor/build-style/style.css');
-  require('EightshiftBlocksStorybookWp/format-library/build-style/style.css');
-  require('EightshiftBlocksStorybookLibsPath/.storybook/parts/editor-styles.scss');
-}
